@@ -1,13 +1,15 @@
 {
-	"name": "nova-public",
-	"version": "4.4.0",
-	"description": "A bot for the Makeshift community and others.",
+	"name": "nova-makeshift",
+	"version": "5.0.0",
+	"description": "A bot for the Makeshift community.",
 	"main": "server.js",
+	"type": "module",
 	"engines": {
-		"node": "10.3.0"
+		"node": "^16.9.0"
 	},
 	"scripts": {
-		"test": "echo \"Error: no test specified\" && exit 1"
+		"test": "standard && node .",
+		"start": "node ."
 	},
 	"keywords": [
 		"discord.js",
@@ -17,19 +19,16 @@
 	],
 	"repository": {
 		"type": "git",
-		"url": "https://github.com/NKN1396/nova-public"
+		"url": "https://github.com/NKN1396/nova-makeshift"
 	},
 	"author": "NKN1396",
 	"license": "MIT",
 	"dependencies": {
-		"common-tags": "^1.8.0",
-		"discord.js": "github:discordjs/discord.js",
-		"discord.js-commando": "github:discordjs/Commando",
-		"discord.js-handles": "github:NKN1396/discord.js-handles#master",
-		"lodash": "^4.17.15",
-		"luxon": "^1.17.2",
-		"request": "^2.88.0",
-		"request-promise-native": "^1.0.8",
-		"striptags": "^3.1.1"
+		"@discordjs/builders": "^0.11.0",
+		"discord.js": "^13.5.1",
+		"lodash": "^4.17.21"
+	},
+	"devDependencies": {
+		"standard": "^16.0.4"
 	}
 }
