@@ -1,58 +1,20 @@
-//External dependencies
-const commando = require("discord.js-commando")
-const discordErrorHandler = require("discord.js-handles")
+// External dependencies
+import { Client, Intents } from 'discord.js'
 
-//Load info
-const token = require("./token.json")	//I'm an idiot - Thanks for the lesson
-const makeshift = require("./src/resources/makeshift.json")
+// Custom dependencies
+import token from './src/resources/token.js' // I'm an idiot, thanks for the lesson
+import registerModules from './src/listeners/index.js'
 
-//Create new Discord client
-const makeshiftbot = new commando.Client({
-	commandPrefix : "!",
-	unknownCommandResponse : false,
-	owner : "153595272465743872"
+const makeshiftbot = new Client({
+  intents: [ Intents.FLAGS.GUILD_MESSAGES, Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_MEMBERS ],
+  partials: [ 'CHANNEL', 'GUILD_MEMBER' ]
 })
 
-//Load and register commands.
-require("./src/commands/index")(makeshiftbot)
+registerModules(makeshiftbot)
 
-let handlerOptions = {
-	logAllGuilds: false,
-	guilds: [
-		makeshift.guild
-	],
-	name: "nova-public",
-	notify: {
-		users: [
-			"153595272465743872"
-		],
-		channels: []
-	}
-}
-
-discordErrorHandler(makeshiftbot, handlerOptions)
-
-//Load custom modules
-require("./src/modules/index")(makeshiftbot)
-
-
-//*
-//Start bot
+// Start bot
 makeshiftbot.login(token)
-	.catch(console.error)
-//*/
-makeshiftbot.on("ready", () => {
-	atNovaHelp(makeshiftbot.user)
-})
-makeshiftbot.on("resume", ()=> {
-	atNovaHelp(makeshiftbot.user)
-})
-
-function atNovaHelp(user){
-	user.setPresence({
-		activity: {
-			name: "@Nova help",
-			type: "PLAYING"
-		}
-	})
-}
+  .catch(console.error)
+  .then(() => {
+    console.log(`Logged in as ${makeshiftbot.user.tag}`)
+  })
