@@ -1,18 +1,15 @@
-const makeshift = require("./../resources/makeshift.json")
+import { GUILD as GUILD_ID, TEXTCHANNELS } from '../resources/makeshift.js'
+const CHANNEL_ID = TEXTCHANNELS.GENERAL
 
-/**
- * Sends a welcome message in #general.
- * @param {*} client The bot client.
- */
-module.exports = function(client){
-	client.on("guildMemberAdd", async guildMember => {
+export default async function (member) {
+  // Check if member joined Makeshift guild
+  const guild = member.guild
+  if (guild.id !== GUILD_ID) return
 
-		//Check if member joined Makeshift guild
-		if(guildMember.guild.id !== makeshift.guild) return
+  const channel = await guild.channels.fetch(CHANNEL_ID)
+    .catch(console.error)
+  if (channel === undefined) { return }
 
-		await client.channels.get(makeshift.channels.text.general).send(`Welcome to the Makeshift clan Discord, ${guildMember.user}!`)
-			.catch(console.error)
-		client.emit("guildMemberGreeted", guildMember)
-
-	})
+  channel.send(`Welcome to the Makeshift clan Discord, ${member}!`)
+    .catch(console.error)
 }
