@@ -1,17 +1,10 @@
-/**
- * Points out obvious sarcasm. Triggered by ending a message with "/s".
- * @param {*} client The Discord.js client to hook into.
- */
-module.exports = function(client){
-	client.on("message", message => {
+const trigger = /(?:^\/s$)|(?:\s\/s$)/g
 
-		if(
-			!message.content.toLowerCase().endsWith(" /s") &&
-			message.content !== "/s"
-		) return
+export default async function (message) {
+  // Check
+  if (message.guild === undefined) { return }
+  if (trigger.test(message.content) === false) { return }
 
-		message.channel.send("*(That was sarcasm)*")
-			.catch(console.error)
-
-	})
+  message.channel.send('*(That was sarcasm)*')
+    .catch(console.error)
 }
